Parte 1/cf-importar-vendas/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub>=2.6.1
google-cloud-storage
pandas
flask
openpyxl